+++ cms/src/features/posts/components/PostCard.tsx
@@ -1,6 +1,8 @@
 import { useParams } from "react-router";
 import NavAnchor from "../../../components/navAnchor/NavAnchor";
 import { Post } from "../../../interfaces/blog";
+import { useModalContext } from "../../../hooks/useModal";
+import ConfirmPostDelete from "./ConfirmPostDelete";
 
 interface Props {
   data: Post;
@@ -9,6 +11,7 @@ interface Props {
 const PostCard: React.FC<Props> = ({ data }) => {
   const { publicId, title, titleSlug, createdAt } = data;
   const { username } = useParams();
+  const { openModal } = useModalContext();
   const postPath = `/${username}/posts/${publicId}/${titleSlug}`;
   const editPostPath = `/${username}/posts/${publicId}/${titleSlug}/edit`;
   // TODO
@@ -20,6 +23,7 @@ const PostCard: React.FC<Props> = ({ data }) => {
       <NavAnchor pathname={postPath} textContent={title} />
       <p>Posted on {createdAt.toString()}</p>
       <NavAnchor pathname={editPostPath} textContent="edit" />
+      <button onClick={() => openModal(<ConfirmPostDelete />)}>Delete</button>
     </li>
   );
 };