+++ cms/src/pages/dashboard/Dashboard.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 
 import { useAuth } from "../../hooks/useAuth";
-
 const Dashboard: React.FC = () => {
   const { accessToken } = useAuth();
+
   return (
     <section>
       <h2>Dashboard</h2>